refactor: Position interface

+++ src/utils.jl
@@ -7,7 +7,7 @@ The Hopcroft-Ullman pairing function, which bijectively maps two positive intege
 
 See https://mathworld.wolfram.com/PairingFunction.html
 """
-@inline pairing(i::Int, j::Int)::Int = div((i+j-2)*(i+j-1), 2) + i
+@inline pairing(i::Integer, j::Integer) = div((i+j-2)*(i+j-1), 2) + i
 
 make_real(x::Number) = SymbolicUtils._isreal(x) ? real(x) : x
